Bundle all JS into a single file
Because we weren't doing this yet, we weren't actually loading the
redirect.js JS code at all.

This also gets rid of some unused JS.
