+++ source/assets/js/components/list-navigation.js
@@ -1,0 +1,10 @@
+$(function() {
+  $(".sl-c-list-navigation-wrapper--collapsible li > ul")
+      .parent()
+      .children("a")
+      .removeAttr("href")
+      .click(function() {
+        $(this).toggleClass("open");
+        return false;
+      });
+});
