--- source/assets/js/components/tables.js
@@ -1,1 +1,0 @@
-$( "sl-c-table" ).wrap( "<div class='sl-c-table-responsive'></div>" );
