--- source/assets/js/styleguide/typography.js
@@ -1,5 +1,0 @@
-$(function(){
-  $("[class*='font-family']").each(function () {
-    $(this).html($(this).css('font-family'))
-  });
-});
