+++ source/layouts/layout.haml
@@ -167,9 +167,5 @@
         },
         debug: false
       });
-    = javascript_include_tag 'vendor/modernizr.custom.min'
-    = javascript_include_tag 'vendor/jquery.min'
-    = javascript_include_tag 'vendor/jquery-ui.min'
-    = javascript_include_tag 'vendor/html5-boilerplate/plugins'
     = javascript_include_tag 'sass'
     = yield_content :js_foot
